requirements.txt
Django
psycopg2
pip
PyJWT
djangorestframework